Refactor dumping of admintools.conf
This is a refactor for the debugDumpAdmintoolsConfg() function. This is
called to dump relevant portions of admintools.conf before and after an
admintools call. Moving this to call to a common wrapper so we don't
have to duplicate the calls so often. This also gets moves it into the
admintools version of the Dispatcher interface.
