+++ pkg/vadmin/add_sc_at.go
@@ -18,7 +18,6 @@ package vadmin
 import (
 	"context"
 
-	"github.com/vertica/vertica-kubernetes/pkg/names"
 	"github.com/vertica/vertica-kubernetes/pkg/vadmin/opts/addsc"
 )
 
@@ -38,6 +37,6 @@ func (a Admintools) AddSubcluster(ctx context.Context, opts ...addsc.Option) err
 		cmd = append(cmd, "--is-secondary")
 	}
 
-	_, _, err := a.PRunner.ExecAdmintools(ctx, s.InitiatorName, names.ServerContainer, cmd...)
+	_, err := a.execAdmintools(ctx, s.InitiatorName, cmd...)
 	return err
 }
