+++ pkg/vadmin/at.go
@@ -17,11 +17,9 @@ package vadmin
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/vertica/vertica-kubernetes/pkg/mgmterrors"
 	"github.com/vertica/vertica-kubernetes/pkg/names"
-	"github.com/vertica/vertica-kubernetes/pkg/paths"
 	"k8s.io/apimachinery/pkg/types"
 	ctrl "sigs.k8s.io/controller-runtime"
 )
@@ -35,15 +33,18 @@ func (a Admintools) logFailure(cmd, genericFailureReason, op string, err error) 
 	return evLogr.LogFailure(cmd, op, err)
 }
 
-// debugDumpAdmintoolsConf will dump specific info from admintools.conf for logging purposes
-// +nolint
-func (a Admintools) debugDumpAdmintoolsConf(ctx context.Context, atPod types.NamespacedName) {
-	// Dump out vital informating from admintools.conf for logging purposes. We
-	// rely on the logging that is done inside ExecInPod.
-	cmd := []string{
-		"bash", "-c",
-		fmt.Sprintf(`ls -l %s && grep '^node\|^v_\|^host' %s`, paths.AdminToolsConf, paths.AdminToolsConf),
+// execAdmintools is a wrapper for admintools tools that handles logging of
+// debug information. The stdout and error of the AT call is returned.
+func (a Admintools) execAdmintools(ctx context.Context, initiatorPod types.NamespacedName, cmd ...string) (string, error) {
+	// Dump relevant contents of the admintools.conf before and after the
+	// admintools calls. We do this for PD purposes to see what changes occurred
+	// in the file.
+	if a.DevMode {
+		a.PRunner.DumpAdmintoolsConf(ctx, initiatorPod)
 	}
-	// Since this is for debugging purposes all errors are ignored
-	a.PRunner.ExecInPod(ctx, atPod, names.ServerContainer, cmd...) //nolint:errcheck
+	stdout, _, err := a.PRunner.ExecAdmintools(ctx, initiatorPod, names.ServerContainer, cmd...)
+	if a.DevMode {
+		a.PRunner.DumpAdmintoolsConf(ctx, initiatorPod)
+	}
+	return stdout, err
 }
